managed-connectivity/community-contributed-connectors/oracle-connector/requirements.txt
google-cloud-dataplex==2.4.0
google-cloud-storage
google-cloud-secret-manager
